refactor(lint): clear the small clang-tidy categories in the relicensed files
Cite the environment-mutating tests and the tiny-AI registry externs
(ADR-0278), band the C-only includes in two shared headers, compare
bit-exact buffers as bytes in the SIMD macro, replace blob pointer
punning with memcpy round-trips, mirror log_generate's roundf, split the
HIP frame feed out of the two ADM border tests, define the Tidy Changed
exclusion filter once and exclude those GPU-only tests from it, route the
SYCL file list through env, and fix four doubled blank lines.

+++ core/test/test_integer_vif_log2.c
@@ -35,7 +35,7 @@
 static void build_log2_table(uint16_t *t)
 {
     for (unsigned i = 0; i < VIF_LOG2_TABLE_SIZE; ++i) {
-        t[i] = (uint16_t)round(log2f((float)(VIF_LOG2_TABLE_OFFSET + i)) * 2048);
+        t[i] = (uint16_t)roundf(log2f((float)(VIF_LOG2_TABLE_OFFSET + i)) * 2048);
     }
 }
